Upgrade whisper and stable-whisper versions, fix humanhash issue, fix AzureService text_offset issue (#53)
* Upgraded whisper and stable-whisper versions

* Replaced playsound with pydub

* Removed humanhash from dependencies

* Fixed issue with AzureService text offset

+++ manim_voiceover/tracker.py
@@ -87,11 +87,11 @@ class VoiceoverTracker:
             )
             self.bookmark_times[mark] = self.start_t + elapsed
 
-    def get_remaining_duration(self, buff: int = 0) -> int:
+    def get_remaining_duration(self, buff: float = 0.) -> float:
         """Returns the remaining duration of the voiceover.
 
         Args:
-            buff (int, optional): A buffer to add to the remaining duration. Defaults to 0.
+            buff (float, optional): A buffer to add to the remaining duration. Defaults to 0.
 
         Returns:
             int: The remaining duration of the voiceover in seconds.
